refactor(ipc): split IpcBridge.cpp into core/ipc + features/ipc
Move the 1720-line monolith into two focused modules:
- src/core/ipc/IpcBridge.cpp (1149 lines): pipe I/O, JSON, session/MAC,
  auth dispatch, tile map, pending events, DispatchCommand, bridge thread
- src/features/ipc/FeatureState.h/cpp (476 lines): feature atomics,
  Apply* functions, all getters/setters, IpcBridge_ApplyFeatureOverrides

Public API (src/ui/IpcBridge.h) and all callers are unchanged.
Add src/core/ipc and src/features/ipc to AdditionalIncludeDirectories.
Builds clean: 0 errors, 0 warnings (Debug x64).
