+++ internal/il2cpp-dll-injection.vcxproj.filters
@@ -10,7 +10,8 @@
     <ClCompile Include="src\core\il2cpp\pch-il2cpp.cpp" />
     <ClCompile Include="src\bootstrap\version.cpp" />
     <ClCompile Include="src\platform\dx11\Dx11.cpp" />
-    <ClCompile Include="src\ui\IpcBridge.cpp" />
+    <ClCompile Include="src\core\ipc\IpcBridge.cpp" />
+    <ClCompile Include="src\features\ipc\FeatureState.cpp" />
     <ClCompile Include="src\ui\Handshake.cpp" />
     <ClCompile Include="src\ui\gui\tabs\CameraTAB.cpp" />
     <ClCompile Include="src\ui\gui\tabs\WorldTAB.cpp" />
@@ -83,6 +84,7 @@
     <ClInclude Include="src\bootstrap\version.h" />
     <ClInclude Include="src\platform\dx11\Dx11.h" />
     <ClInclude Include="src\ui\IpcBridge.h" />
+    <ClInclude Include="src\features\ipc\FeatureState.h" />
     <ClInclude Include="src\ui\Handshake.h" />
     <ClInclude Include="src\ui\gui\tabs\CameraTAB.h" />
     <ClInclude Include="src\ui\gui\tabs\WorldTAB.h" />
