+++ internal/src/core/ipc/IpcBridge.cpp
@@ -1,0 +1,1149 @@
+#include "pch-il2cpp.h"
+#include "IpcBridge.h"
+#include "FeatureState.h"
+#include "Handshake.h"
+#include "xorstr.h"
+#include "main.h"
+#include "settings.h"
+#include "DbgFileLog.h"
+
+#if __has_include("BuildSecrets.h")
+#include "BuildSecrets.h"
+#endif
+#ifndef BUILD_PIPE_NAME
+#define BUILD_PIPE_NAME "\\\\.\\pipe\\lfg-dev-bridge"
+#endif
+
+#include "LocalPlayer.h"
+#include "GameState.h"
+#include "RuntimeOffsets.h"
+#include "AutoAim.h"
+#include "ProjNoclip.h"
+#include "FpsSetter.h"
+#include "Noclip.h"
+#include "GhostHit.h"
+#include "SkinChanger.h"
+#include "gui/tabs/TestTAB.h"
+#include "gui/tabs/CameraTAB.h"
+#include "gui/tabs/CombatTab/CombatTAB.h"
+#include "DangerPlanner.h"
+#include "XDodge.h"
+#include "RolloutDodge.h"
+#include "SpeedHack.h"
+
+#include <cstdio>
+#include <cstring>
+#include <cmath>
+#include <cstdlib>
+#include <cstdint>
+#include <climits>
+#include <limits>
+#include <algorithm>
+#include <cctype>
+#include <unordered_map>
+#include <unordered_set>
+#include <vector>
+#include <mutex>
+#include <atomic>
+#include <string>
+#include <chrono>
+#include <iostream>
+
+#ifdef _DEBUG
+#define DbgLog(fmt, ...) do { \
+    char _b[512]; snprintf(_b, sizeof(_b), fmt, ##__VA_ARGS__); \
+    std::cout << "[IpcBridge] " << _b << std::endl; \
+} while(0)
+#else
+#define DbgLog(fmt, ...) (void)0
+#endif
+
+static void DebugSessionLogSetFeature(const char* key, const char* valueType)
+{
+#ifdef _DEBUG
+    FILE* f = nullptr;
+    fopen_s(&f, "/home/jesse/LFG/.cursor/debug-8c94fe.log", "ab");
+    if (!f) return;
+    const auto ts = std::chrono::duration_cast<std::chrono::milliseconds>(
+        std::chrono::system_clock::now().time_since_epoch()).count();
+    fprintf(
+        f,
+        "{\"sessionId\":\"8c94fe\",\"runId\":\"feature-pipeline-debug\",\"hypothesisId\":\"H4\","
+        "\"location\":\"src/core/ipc/IpcBridge.cpp:setFeature\",\"message\":\"DLL setFeature applied\","
+        "\"data\":{\"key\":\"%s\",\"valueType\":\"%s\"},\"timestamp\":%lld}\n",
+        key ? key : "",
+        valueType ? valueType : "",
+        static_cast<long long>(ts)
+    );
+    fclose(f);
+#else
+    (void)key; (void)valueType;
+#endif
+}
+
+static constexpr int DLL_BUILD = 2;
+
+// ── Pipe constants ────────────────────────────────────────────────────────────
+
+static const char* PipeName()        { return BUILD_PIPE_NAME; }
+static const DWORD PIPE_BUFFER_SIZE = 65536;
+
+// ── Tile map (from bot-client tileUpdate / noWalkInit packets) ────────────────
+// Key: (x & 0xFFFF) << 16 | (y & 0xFFFF).  Value: tile type.
+
+static std::unordered_map<uint32_t, uint16_t> s_tileMap;
+static std::unordered_set<uint16_t>           s_noWalkTypes;
+static std::mutex                              s_tileMutex;
+
+bool IpcBridge_IsTileWalkable(float worldX, float worldY)
+{
+    int tx = (int)floorf(worldX);
+    int ty = (int)floorf(worldY);
+    if (tx < 0 || ty < 0 || tx > 0xFFFF || ty > 0xFFFF) return true;
+    uint32_t key = (static_cast<uint32_t>(tx & 0xFFFF) << 16)
+                 |  static_cast<uint32_t>(ty & 0xFFFF);
+    std::lock_guard<std::mutex> lk(s_tileMutex);
+    auto it = s_tileMap.find(key);
+    if (it == s_tileMap.end()) return true;
+    return s_noWalkTypes.find(it->second) == s_noWalkTypes.end();
+}
+
+void IpcBridge_GetTileStats(int* outTileCount, int* outNoWalkTypeCount)
+{
+    std::lock_guard<std::mutex> lk(s_tileMutex);
+    if (outTileCount)       *outTileCount       = static_cast<int>(s_tileMap.size());
+    if (outNoWalkTypeCount) *outNoWalkTypeCount  = static_cast<int>(s_noWalkTypes.size());
+}
+
+int IpcBridge_CopyUniqueTypeEntries(IpcTileTypeEntry* buf, int maxCount)
+{
+    if (!buf || maxCount <= 0) return 0;
+    std::lock_guard<std::mutex> lk(s_tileMutex);
+
+    std::unordered_map<uint16_t, int> typeCounts;
+    typeCounts.reserve(s_tileMap.size() / 4 + 1);
+    for (const auto& kv : s_tileMap)
+        typeCounts[kv.second]++;
+
+    int n = 0;
+    for (const auto& tc : typeCounts) {
+        if (n >= maxCount) break;
+        buf[n].typeId = tc.first;
+        buf[n].count  = tc.second;
+        buf[n].noWalk = (s_noWalkTypes.find(tc.first) != s_noWalkTypes.end());
+        ++n;
+    }
+    std::sort(buf, buf + n, [](const IpcTileTypeEntry& a, const IpcTileTypeEntry& b) {
+        return a.typeId < b.typeId;
+    });
+    return n;
+}
+
+// ── Auth state ────────────────────────────────────────────────────────────────
+
+static Handshake::AuthState s_auth = {};
+
+const char* IpcBridge_GetUserId()       { return s_auth.userId; }
+bool        IpcBridge_IsAuthenticated() { return Handshake::IsHealthy(&s_auth); }
+
+// ── Shutdown flag ─────────────────────────────────────────────────────────────
+
+static std::atomic<bool> s_shutdown{false};
+
+void IpcBridge_RequestShutdown() { s_shutdown = true; }
+
+// ── Pending events from worker threads (e.g. GhostHit) ────────────────────────
+
+struct PendingEvent {
+    char pluginId[32];
+    char action[128];
+};
+
+static std::mutex                s_pendingEventsMutex;
+static std::vector<PendingEvent> s_pendingEvents;
+static constexpr size_t          kPendingEventsCap = 64;
+
+void IpcBridge_EmitPredictedHit(int ownerObjId, int bulletId)
+{
+    PendingEvent ev{};
+    std::snprintf(ev.pluginId, sizeof(ev.pluginId), "%s", "ghostHit");
+    std::snprintf(ev.action,   sizeof(ev.action),   "%d:%d", ownerObjId, bulletId);
+
+    std::lock_guard<std::mutex> lk(s_pendingEventsMutex);
+    if (s_pendingEvents.size() < kPendingEventsCap)
+        s_pendingEvents.push_back(ev);
+}
+
+// ── Length-prefixed message I/O ───────────────────────────────────────────────
+
+static bool PipeWriteMessage(HANDLE hPipe, const char* json, int len)
+{
+    uint32_t netLen = static_cast<uint32_t>(len);
+    DWORD written = 0;
+    if (!WriteFile(hPipe, &netLen, 4, &written, NULL) || written != 4) return false;
+    if (!WriteFile(hPipe, json, netLen, &written, NULL) || written != netLen) return false;
+    return true;
+}
+
+static int PipeReadMessage(HANDLE hPipe, char* buf, int bufSize)
+{
+    DWORD bytesAvail = 0;
+    if (!PeekNamedPipe(hPipe, NULL, 0, NULL, &bytesAvail, NULL)) return -1;
+    if (bytesAvail < 4) return 0;
+
+    uint32_t msgLen = 0;
+    DWORD bytesRead = 0;
+    if (!ReadFile(hPipe, &msgLen, 4, &bytesRead, NULL) || bytesRead != 4) return -1;
+    if (msgLen == 0 || msgLen >= (uint32_t)bufSize) return -1;
+
+    if (!ReadFile(hPipe, buf, msgLen, &bytesRead, NULL) || bytesRead != msgLen) return -1;
+    buf[msgLen] = '\0';
+    return (int)msgLen;
+}
+
+// ── JSON builders ─────────────────────────────────────────────────────────────
+
+static int BuildHelloJson(char* buf, int bufSize, const char* challenge)
+{
+    return snprintf(buf, bufSize,
+        "{\"type\":\"hello\",\"version\":3,\"protocol\":\"bridge-v3\","
+        "\"build\":%d,\"challenge\":\"%s\",\"features\":["
+        "\"autoDodge\",\"autoAim\",\"tileMap\""
+        "]}", DLL_BUILD, challenge);
+}
+
+static int BuildAuthResultJson(char* buf, int bufSize, bool ok, const char* response)
+{
+    if (ok)
+        return snprintf(buf, bufSize, "{\"type\":\"authResult\",\"ok\":true,\"response\":\"%s\"}", response);
+    else
+        return snprintf(buf, bufSize, "{\"type\":\"authResult\",\"ok\":false}");
+}
+
+static int BuildHeartbeatJson(char* buf, int bufSize, const char* nonce, uint64_t seq, const char* mac)
+{
+    return snprintf(buf, bufSize,
+        "{\"type\":\"heartbeat\",\"nonce\":\"%s\",\"seq\":\"%llu\",\"mac\":\"%s\"}",
+        nonce, static_cast<unsigned long long>(seq), mac);
+}
+
+static int BuildHeartbeatRespJson(char* buf, int bufSize, const char* response, uint64_t seq, const char* mac)
+{
+    return snprintf(buf, bufSize,
+        "{\"type\":\"heartbeatResp\",\"response\":\"%s\",\"seq\":\"%llu\",\"mac\":\"%s\"}",
+        response, static_cast<unsigned long long>(seq), mac);
+}
+
+static int BuildUnresolvedClassesJson(char* buf, int bufSize, const char* classes, uint64_t seq, const char* mac)
+{
+    return snprintf(buf, bufSize,
+        "{\"type\":\"unresolvedClasses\",\"classes\":\"%s\",\"seq\":\"%llu\",\"mac\":\"%s\"}",
+        classes, static_cast<unsigned long long>(seq), mac);
+}
+
+static int BuildPlayerJson(char* buf, int bufSize, uint64_t seq, const char* mac)
+{
+    float   posX  = LocalPlayer::GetX();
+    float   posY  = LocalPlayer::GetY();
+    int32_t hp    = LocalPlayer::GetHP();
+    int32_t maxHp = LocalPlayer::GetMaxHP();
+    int32_t def   = LocalPlayer::GetDefense();
+
+    if (!LocalPlayer::GetPtr())
+        return snprintf(buf, bufSize,
+            "{\"type\":\"player\",\"alive\":false,\"seq\":\"%llu\",\"mac\":\"%s\"}",
+            static_cast<unsigned long long>(seq), mac);
+
+    return snprintf(buf, bufSize,
+        "{\"type\":\"player\",\"alive\":true,"
+        "\"hp\":%d,\"maxHp\":%d,\"def\":%d,"
+        "\"posX\":%.3f,\"posY\":%.3f,\"seq\":\"%llu\",\"mac\":\"%s\"}",
+        hp, maxHp, def, (double)posX, (double)posY,
+        static_cast<unsigned long long>(seq), mac);
+}
+
+static int BuildHotkeyEventJson(
+    char* buf, int bufSize,
+    const char* pluginId, const char* action,
+    bool value, uint64_t seq, const char* mac)
+{
+    return snprintf(buf, bufSize,
+        "{\"type\":\"hotkeyEvent\",\"pluginId\":\"%s\",\"action\":\"%s\","
+        "\"value\":%s,\"seq\":\"%llu\",\"mac\":\"%s\"}",
+        pluginId, action, value ? "true" : "false",
+        static_cast<unsigned long long>(seq), mac);
+}
+
+// ── Minimal JSON helpers ──────────────────────────────────────────────────────
+
+static char* JsonGetString(char* json, const char* key, char* valBuf, int valBufSize)
+{
+    char pattern[128];
+    snprintf(pattern, sizeof(pattern), "\"%s\":\"", key);
+    char* start = strstr(json, pattern);
+    if (!start) return NULL;
+    start += strlen(pattern);
+    char* end = strchr(start, '"');
+    if (!end) return NULL;
+    int len = (int)(end - start);
+    if (len >= valBufSize) len = valBufSize - 1;
+    memcpy(valBuf, start, len);
+    valBuf[len] = '\0';
+    return valBuf;
+}
+
+static bool IsAsciiIdSafe(const char* s)
+{
+    if (!s || !*s) return false;
+    size_t len = strlen(s);
+    if (len > 96) return false;
+    for (size_t i = 0; i < len; ++i) {
+        const char c = s[i];
+        const bool ok =
+            (c >= 'a' && c <= 'z') || (c >= 'A' && c <= 'Z') ||
+            (c >= '0' && c <= '9') || c == '-' || c == '_' || c == '.';
+        if (!ok) return false;
+    }
+    return true;
+}
+
+static bool ParseUint64Dec(const char* s, uint64_t* out)
+{
+    if (!s || !*s || !out) return false;
+    uint64_t acc = 0;
+    for (const char* p = s; *p; ++p) {
+        if (*p < '0' || *p > '9') return false;
+        const uint64_t digit = static_cast<uint64_t>(*p - '0');
+        const uint64_t next  = acc * 10ULL + digit;
+        if (next < acc) return false;
+        acc = next;
+    }
+    *out = acc;
+    return true;
+}
+
+static bool ConstantTimeHexEq64(const char* a, const char* b)
+{
+    if (!a || !b) return false;
+    if (strlen(a) != 64 || strlen(b) != 64) return false;
+    uint8_t diff = 0;
+    for (int i = 0; i < 64; ++i) diff |= static_cast<uint8_t>(a[i] ^ b[i]);
+    return diff == 0;
+}
+
+// ── Session key + MAC helpers ─────────────────────────────────────────────────
+
+static bool DeriveSessionKey(
+    const char* serverChallenge, const char* clientChallenge,
+    const char* userId, const char* clientPid,
+    uint8_t outKey[32])
+{
+    if (!serverChallenge || !clientChallenge || !userId || !clientPid || !outKey) return false;
+    if (strlen(serverChallenge) != 64 || !Handshake::IsHexString(serverChallenge, 64)) return false;
+    if (strlen(clientChallenge) != 64 || !Handshake::IsHexString(clientChallenge, 64)) return false;
+    if (!IsAsciiIdSafe(userId)) return false;
+    uint64_t pidNum = 0;
+    if (!ParseUint64Dec(clientPid, &pidNum) || pidNum == 0) return false;
+
+    char data[512] = {};
+    snprintf(data, sizeof(data), "%s|%s|%s|%s|%s|session-v2",
+        serverChallenge, clientChallenge, userId, clientPid, PipeName());
+    return Handshake::HmacSha256(
+        Handshake::GetSharedKey(), 32,
+        reinterpret_cast<const uint8_t*>(data), strlen(data),
+        outKey);
+}
+
+static bool ComputeSessionMacHex(
+    const uint8_t key[32], uint64_t seq,
+    const char* type, const char* payload,
+    char outHex[65])
+{
+    if (!key || !type || !payload || !outHex) return false;
+    char data[1024] = {};
+    snprintf(data, sizeof(data), "%llu|%s|%s", static_cast<unsigned long long>(seq), type, payload);
+    uint8_t mac[32] = {};
+    if (!Handshake::HmacSha256(key, 32,
+            reinterpret_cast<const uint8_t*>(data), strlen(data), mac))
+        return false;
+    Handshake::ToHex(mac, 32, outHex);
+    return true;
+}
+
+static bool WriteSignedHotkeyEvent(
+    HANDLE hPipe, char* msgBuf, int msgBufSize,
+    const char* pluginId, const char* action, bool value)
+{
+    if (!hPipe || !msgBuf || !pluginId || !action) return false;
+    char payload[128] = {};
+    snprintf(payload, sizeof(payload), "%s|%s|%s", pluginId, action, value ? "true" : "false");
+
+    const uint64_t outSeq = s_auth.nextServerSeq++;
+    char outMac[65] = {};
+    if (!ComputeSessionMacHex(s_auth.sessionKey, outSeq, "hotkeyEvent", payload, outMac))
+        return false;
+
+    const int len = BuildHotkeyEventJson(msgBuf, msgBufSize, pluginId, action, value, outSeq, outMac);
+    return PipeWriteMessage(hPipe, msgBuf, len);
+}
+
+static void BuildPlayerSigPayload(char* outBuf, int outBufSize)
+{
+    if (!outBuf || outBufSize <= 0) return;
+    float   posX  = LocalPlayer::GetX();
+    float   posY  = LocalPlayer::GetY();
+    int32_t hp    = LocalPlayer::GetHP();
+    int32_t maxHp = LocalPlayer::GetMaxHP();
+    int32_t def   = LocalPlayer::GetDefense();
+
+    if (!LocalPlayer::GetPtr()) {
+        snprintf(outBuf, outBufSize, "alive:false");
+        return;
+    }
+    snprintf(outBuf, outBufSize,
+        "alive:true|hp:%d|maxHp:%d|posX:%.3f|posY:%.3f|def:%d",
+        hp, maxHp, (double)posX, (double)posY, def);
+}
+
+static bool VerifyClientSeqAndMac(const char* seqStr, const char* macHex, const char* type, const char* payload)
+{
+    (void)macHex; (void)type; (void)payload;
+    if (!s_auth.authenticated) return false;
+    if (!seqStr) return false;
+    uint64_t seq = 0;
+    if (!ParseUint64Dec(seqStr, &seq)) return false;
+    if (seq <= s_auth.lastClientSeq) return false;
+    s_auth.lastClientSeq = seq;
+    return true;
+}
+
+// ── Auth/heartbeat dispatcher ─────────────────────────────────────────────────
+
+static bool DispatchAuthMessage(char* json, HANDLE hPipe, char* msgBuf, int msgBufSize)
+{
+    char typeBuf[64] = {};
+    if (!JsonGetString(json, "type", typeBuf, sizeof(typeBuf))) return false;
+
+    if (strcmp(typeBuf, "auth") == 0)
+    {
+        char userId[128]          = {};
+        char response[128]        = {};
+        char clientChallenge[128] = {};
+        char protocol[32]         = {};
+        char clientPid[32]        = {};
+
+        if (!JsonGetString(json, "userId",     userId,           sizeof(userId))          ||
+            !JsonGetString(json, "response",   response,         sizeof(response))         ||
+            !JsonGetString(json, "challenge",  clientChallenge,  sizeof(clientChallenge))  ||
+            !JsonGetString(json, "protocol",   protocol,         sizeof(protocol))          ||
+            !JsonGetString(json, "clientPid",  clientPid,        sizeof(clientPid)))
+        {
+            DbgLog("Auth message missing required fields.");
+            int len = BuildAuthResultJson(msgBuf, msgBufSize, false, "");
+            PipeWriteMessage(hPipe, msgBuf, len);
+            return true;
+        }
+
+        if (!IsAsciiIdSafe(userId) ||
+            strcmp(protocol, "bridge-v3") != 0 ||
+            strlen(response) != 64 || !Handshake::IsHexString(response, 64) ||
+            strlen(clientChallenge) != 64 || !Handshake::IsHexString(clientChallenge, 64) ||
+            strlen(s_auth.pendingChallenge) != 64 || !Handshake::IsHexString(s_auth.pendingChallenge, 64))
+        {
+            DbgLog("Auth payload failed format validation.");
+            int len = BuildAuthResultJson(msgBuf, msgBufSize, false, "");
+            PipeWriteMessage(hPipe, msgBuf, len);
+            return true;
+        }
+
+        strncpy_s(s_auth.userId, sizeof(s_auth.userId), userId, _TRUNCATE);
+        if (!DeriveSessionKey(s_auth.pendingChallenge, clientChallenge, s_auth.userId, clientPid, s_auth.sessionKey))
+        {
+            DbgLog("Session key derivation failed.");
+            int len = BuildAuthResultJson(msgBuf, msgBufSize, false, "");
+            PipeWriteMessage(hPipe, msgBuf, len);
+            return true;
+        }
+        s_auth.authenticated     = true;
+        s_auth.sessionReady      = true;
+        s_auth.heartbeatMisses   = 0;
+        s_auth.lastHeartbeatRecv = GetTickCount64();
+        s_auth.lastClientSeq     = 0;
+        s_auth.nextServerSeq     = 1;
+
+        DbgLog("Auth OK: userId=%s", userId);
+
+        char dllResponse[65] = {};
+        if (!Handshake::ComputeResponse(clientChallenge, strlen(clientChallenge), dllResponse))
+        {
+            DbgLog("Auth response generation failed.");
+            int len = BuildAuthResultJson(msgBuf, msgBufSize, false, "");
+            PipeWriteMessage(hPipe, msgBuf, len);
+            return true;
+        }
+        int len = BuildAuthResultJson(msgBuf, msgBufSize, true, dllResponse);
+        PipeWriteMessage(hPipe, msgBuf, len);
+        return true;
+    }
+
+    if (strcmp(typeBuf, "heartbeatResp") == 0)
+    {
+        char response[128] = {};
+        char seqStr[32]    = {};
+        char macHex[128]   = {};
+        if (!JsonGetString(json, "response", response, sizeof(response))) return true;
+        if (!JsonGetString(json, "seq",      seqStr,   sizeof(seqStr)))   return true;
+        if (!JsonGetString(json, "mac",      macHex,   sizeof(macHex)))   return true;
+        if (strlen(response) != 64 || !Handshake::IsHexString(response, 64)) {
+            s_auth.heartbeatMisses++;
+            return true;
+        }
+        if (!VerifyClientSeqAndMac(seqStr, macHex, "heartbeatResp", response)) {
+            s_auth.heartbeatMisses++;
+            return true;
+        }
+
+        if (s_auth.challengePending &&
+            Handshake::VerifyResponse(s_auth.pendingChallenge, strlen(s_auth.pendingChallenge), response))
+        {
+            s_auth.challengePending  = false;
+            s_auth.heartbeatMisses   = 0;
+            s_auth.lastHeartbeatRecv = GetTickCount64();
+        }
+        else
+        {
+            s_auth.heartbeatMisses++;
+        }
+        return true;
+    }
+
+    if (strcmp(typeBuf, "heartbeat") == 0)
+    {
+        char nonce[128]  = {};
+        char seqStr[32]  = {};
+        char macHex[128] = {};
+        if (!JsonGetString(json, "nonce", nonce,  sizeof(nonce)))  return true;
+        if (!JsonGetString(json, "seq",   seqStr, sizeof(seqStr))) return true;
+        if (!JsonGetString(json, "mac",   macHex, sizeof(macHex))) return true;
+        if (strlen(nonce) != 64 || !Handshake::IsHexString(nonce, 64)) return true;
+        if (!VerifyClientSeqAndMac(seqStr, macHex, "heartbeat", nonce)) return true;
+        char resp[65] = {};
+        if (!Handshake::ComputeResponse(nonce, strlen(nonce), resp)) return true;
+        const uint64_t outSeq = s_auth.nextServerSeq++;
+        char outMac[65] = {};
+        if (!ComputeSessionMacHex(s_auth.sessionKey, outSeq, "heartbeatResp", resp, outMac)) return true;
+        int len = BuildHeartbeatRespJson(msgBuf, msgBufSize, resp, outSeq, outMac);
+        PipeWriteMessage(hPipe, msgBuf, len);
+        return true;
+    }
+
+    return false;
+}
+
+// ── JSON bool/number helpers ──────────────────────────────────────────────────
+
+static bool JsonGetBool(char* json, const char* key)
+{
+    char valBuf[16] = {};
+    if (JsonGetString(json, key, valBuf, sizeof(valBuf)))
+        return strcmp(valBuf, "true") == 0 || strcmp(valBuf, "1") == 0;
+
+    char pattern[128];
+    snprintf(pattern, sizeof(pattern), "\"%s\":", key);
+    const char* p = strstr(json, pattern);
+    if (!p) return false;
+    p += strlen(pattern);
+    while (*p == ' ') p++;
+    return strncmp(p, "true", 4) == 0;
+}
+
+static bool JsonGetNumberToken(char* json, const char* key, char* outBuf, int outBufSize)
+{
+    if (!json || !key || !outBuf || outBufSize <= 1) return false;
+    char pattern[128];
+    snprintf(pattern, sizeof(pattern), "\"%s\":", key);
+    const char* p = strstr(json, pattern);
+    if (!p) return false;
+    p += strlen(pattern);
+    while (*p == ' ') p++;
+
+    int i = 0;
+    if (*p == '-') {
+        if (i < outBufSize - 1) outBuf[i++] = *p;
+        ++p;
+    }
+    bool seenDigit = false;
+    while ((*p >= '0' && *p <= '9') || *p == '.') {
+        seenDigit = true;
+        if (i >= outBufSize - 1) return false;
+        outBuf[i++] = *p++;
+    }
+    if (!seenDigit) return false;
+    outBuf[i] = '\0';
+    return true;
+}
+
+// ── Hotkey string → VK resolver ──────────────────────────────────────────────
+
+namespace {
+
+static int ResolveHotkeyVk(const char* raw)
+{
+    if (!raw) return 0;
+
+    std::string key;
+    for (const char* p = raw; *p; ++p) {
+        const unsigned char ch = static_cast<unsigned char>(*p);
+        if (!std::isspace(ch))
+            key.push_back(static_cast<char>(std::toupper(ch)));
+    }
+    if (key.empty() || key == "NONE" || key == "OFF") return 0;
+    if (key.rfind("VK_", 0) == 0) key.erase(0, 3);
+
+    if (key.size() == 1) {
+        const char ch = key[0];
+        if ((ch >= 'A' && ch <= 'Z') || (ch >= '0' && ch <= '9'))
+            return static_cast<int>(ch);
+    }
+
+    if (key.size() >= 2 && key[0] == 'F') {
+        const int fn = atoi(key.c_str() + 1);
+        if (fn >= 1 && fn <= 24) return VK_F1 + (fn - 1);
+    }
+
+    if (key == "SPACE")                       return VK_SPACE;
+    if (key == "TAB")                         return VK_TAB;
+    if (key == "ESC"  || key == "ESCAPE")     return VK_ESCAPE;
+    if (key == "SHIFT")                       return VK_SHIFT;
+    if (key == "CTRL" || key == "CONTROL")    return VK_CONTROL;
+    if (key == "ALT"  || key == "MENU")       return VK_MENU;
+    if (key == "INSERT" || key == "INS")      return VK_INSERT;
+    if (key == "DELETE" || key == "DEL")      return VK_DELETE;
+    if (key == "HOME")                        return VK_HOME;
+    if (key == "END")                         return VK_END;
+    if (key == "PAGEUP"   || key == "PGUP")   return VK_PRIOR;
+    if (key == "PAGEDOWN" || key == "PGDN")   return VK_NEXT;
+    if (key == "UP")                          return VK_UP;
+    if (key == "DOWN")                        return VK_DOWN;
+    if (key == "LEFT")                        return VK_LEFT;
+    if (key == "RIGHT")                       return VK_RIGHT;
+
+    if (key.rfind("NUMPAD", 0) == 0) {
+        const int n = atoi(key.c_str() + 6);
+        if (n >= 0 && n <= 9) return VK_NUMPAD0 + n;
+    }
+
+    return 0;
+}
+
+} // namespace
+
+// ── Command dispatcher ────────────────────────────────────────────────────────
+
+static void DispatchCommand(char* json)
+{
+    char typeBuf[64]  = {};
+    char seqStr[32]   = {};
+    char macHex[128]  = {};
+    if (!JsonGetString(json, "type", typeBuf, sizeof(typeBuf))) return;
+    if (!JsonGetString(json, "seq",  seqStr,  sizeof(seqStr)))  return;
+    if (!JsonGetString(json, "mac",  macHex,  sizeof(macHex)))  return;
+
+    if (strcmp(typeBuf, "clearTiles") == 0)
+    {
+        if (!VerifyClientSeqAndMac(seqStr, macHex, "clearTiles", "")) return;
+        std::lock_guard<std::mutex> lk(s_tileMutex);
+        s_tileMap.clear();
+        DbgLog("Tile map cleared.");
+    }
+    else if (strcmp(typeBuf, "noWalkInit") == 0)
+    {
+        char typesBuf[8192] = {};
+        if (!JsonGetString(json, "types", typesBuf, sizeof(typesBuf))) return;
+        if (!VerifyClientSeqAndMac(seqStr, macHex, "noWalkInit", typesBuf)) return;
+        std::lock_guard<std::mutex> lk(s_tileMutex);
+        s_noWalkTypes.clear();
+        char* saveptr = typesBuf;
+        char* tok = strtok_s(saveptr, " ", &saveptr);
+        int count = 0;
+        while (tok) {
+            int t = atoi(tok);
+            if (t > 0 && t <= 0xFFFF) { s_noWalkTypes.insert(static_cast<uint16_t>(t)); ++count; }
+            tok = strtok_s(nullptr, " ", &saveptr);
+        }
+        DbgLog("noWalkInit: %d impassable tile types.", count);
+    }
+    else if (strcmp(typeBuf, "tileUpdate") == 0)
+    {
+        char tilesBuf[65000] = {};
+        if (!JsonGetString(json, "tiles", tilesBuf, sizeof(tilesBuf))) return;
+        if (!VerifyClientSeqAndMac(seqStr, macHex, "tileUpdate", tilesBuf)) return;
+        std::lock_guard<std::mutex> lk(s_tileMutex);
+        char* saveptr = tilesBuf;
+        char* tok = strtok_s(saveptr, " ", &saveptr);
+        while (tok) {
+            int x = 0, y = 0, t = 0;
+            if (sscanf_s(tok, "%d:%d:%d", &x, &y, &t) == 3
+                && x >= 0 && x <= 0xFFFF && y >= 0 && y <= 0xFFFF)
+            {
+                uint32_t key = (static_cast<uint32_t>(x & 0xFFFF) << 16)
+                             |  static_cast<uint32_t>(y & 0xFFFF);
+                s_tileMap[key] = static_cast<uint16_t>(t);
+            }
+            tok = strtok_s(nullptr, " ", &saveptr);
+        }
+    }
+    else if (strcmp(typeBuf, "setFeature") == 0)
+    {
+        char keyBuf[64]    = {};
+        char valueType[8]  = {};
+        char valueNorm[128]= {};
+        if (!JsonGetString(json, "key",       keyBuf,    sizeof(keyBuf)))    return;
+        if (!JsonGetString(json, "valueType", valueType, sizeof(valueType))) return;
+
+        if (strcmp(valueType, "b") == 0) {
+            bool enabled = JsonGetBool(json, "value");
+            strncpy_s(valueNorm, sizeof(valueNorm), enabled ? "true" : "false", _TRUNCATE);
+        } else if (strcmp(valueType, "n") == 0) {
+            if (!JsonGetNumberToken(json, "value", valueNorm, sizeof(valueNorm))) return;
+        } else if (strcmp(valueType, "s") == 0) {
+            if (!JsonGetString(json, "value", valueNorm, sizeof(valueNorm))) return;
+        } else {
+            return;
+        }
+
+        char payload[256] = {};
+        snprintf(payload, sizeof(payload), "%s|%s|%s", keyBuf, valueType, valueNorm);
+        if (!VerifyClientSeqAndMac(seqStr, macHex, "setFeature", payload)) {
+            DBG_FILE_LOG("[IpcBridge] setFeature HMAC REJECTED: key=" << keyBuf
+                << " valueType=" << valueType << " value=" << valueNorm);
+            return;
+        }
+        DBG_FILE_LOG("[IpcBridge] setFeature: key=" << keyBuf
+            << " valueType=" << valueType << " value=" << valueNorm);
+        DebugSessionLogSetFeature(keyBuf, valueType);
+
+        if (strcmp(keyBuf, "overlayEnabled") == 0)
+        {
+            bool enabled = JsonGetBool(json, "value");
+            s_overlayEnabled.store(enabled, std::memory_order_relaxed);
+            if (!enabled) settings.bShowMenu = false;
+            DbgLog("overlayEnabled = %s", enabled ? "true" : "false");
+        } else if (strcmp(keyBuf, "internalUnloadDll") == 0) {
+            if (JsonGetBool(json, "value")) {
+                DBG_FILE_LOG("[IpcBridge] internalUnloadDll requested.");
+                IpcBridge_RequestShutdown();
+                if (hUnloadEvent) SetEvent(hUnloadEvent);
+            }
+        } else if (strcmp(keyBuf, "autoAimEnabled") == 0) {
+            IpcBridge_SetAutoAimEnabled(JsonGetBool(json, "value"));
+        } else if (strcmp(keyBuf, "autoAimMode") == 0) {
+            IpcBridge_SetAutoAimMode(atoi(valueNorm));
+        } else if (strcmp(keyBuf, "autoAimPrioritizeBosses") == 0) {
+            AutoAim::SetPrioritizeBosses(JsonGetBool(json, "value"));
+        } else if (strcmp(keyBuf, "autoAimIgnoreWalls") == 0) {
+            AutoAim::SetIgnoreWalls(JsonGetBool(json, "value"));
+        } else if (strcmp(keyBuf, "projectileNoclipEnabled") == 0) {
+            const bool enabled = JsonGetBool(json, "value");
+            s_featProjectileNoclipEnabled.store(enabled ? 1 : 0, std::memory_order_relaxed);
+            if (!enabled) ProjNoclip::SetEnabled(false);
+        } else if (strcmp(keyBuf, "clientDefense") == 0) {
+            s_featClientDefense.store(static_cast<int32_t>(atoi(valueNorm)), std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "clientClassType") == 0) {
+            s_featClientClassType.store(static_cast<int32_t>(atoi(valueNorm)), std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "autoDodgeMode") == 0) {
+            IpcBridge_SetAutoDodgeMode(atoi(valueNorm));
+        } else if (strcmp(keyBuf, "autoDodgeHorizonMs") == 0) {
+            IpcBridge_SetAutoDodgeHorizonMs(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "autoDodgeHitboxPadding") == 0) {
+            IpcBridge_SetAutoDodgeHitboxPadding(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "autoDodgeWallAvoid") == 0) {
+            IpcBridge_SetAutoDodgeWallAvoid(JsonGetBool(json, "value"));
+        // ── XDodge settings ──────────────────────────────────────────────────
+        } else if (strcmp(keyBuf, "xdodgeHitScale") == 0) {
+            XDodge::SetHitScale(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "xdodgeRebuildN") == 0) {
+            XDodge::SetRebuildN(static_cast<int>(atoi(valueNorm)));
+        } else if (strcmp(keyBuf, "xdodgePlanStepMs") == 0) {
+            XDodge::SetPlanStepMs(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "xdodgeSearchRadius") == 0) {
+            XDodge::SetSearchRadius(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "xdodgeAstar") == 0) {
+            XDodge::SetAstarEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeWeighting") == 0) {
+            XDodge::SetWeightingEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeSmartGoal") == 0) {
+            XDodge::SetSmartGoalEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgePerpBias") == 0) {
+            XDodge::SetPerpEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeSpeedMatch") == 0) {
+            XDodge::SetSpeedMatchEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeLockFollow") == 0) {
+            DangerPlanner::SetLockFollowEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeAutoLock") == 0) {
+            DangerPlanner::SetAutoLockMode(atoi(valueNorm));
+        } else if (strcmp(keyBuf, "xdodgeWalkCache") == 0) {
+            XDodge::SetWalkCacheEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeWallAvoid") == 0) {
+            XDodge::SetWallAvoidEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeArbiter") == 0) {
+            XDodge::SetArbiterEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeBfsBias") == 0) {
+            XDodge::SetBfsBiasEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeCcd") == 0) {
+            XDodge::SetCcdEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeCcdPad") == 0) {
+            XDodge::SetCcdPad(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "xdodgeCatalog") == 0) {
+            XDodge::SetCatalogEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeNotifyHit") == 0) {
+            XDodge::OnPlayerHit();
+        } else if (strcmp(keyBuf, "xdodgeDrawPath") == 0) {
+            XDodge::SetDrawPathEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeDrawProjPred") == 0) {
+            XDodge::SetDrawProjPredEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeDebugPredLongMs") == 0) {
+            XDodge::SetDebugPredLongMs(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "xdodgeAvoidEnemies") == 0) {
+            XDodge::SetAvoidEnemiesEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeGhostHit") == 0) {
+            GhostHit::SetEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeLosGoal") == 0) {
+            XDodge::SetLosGoalEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeWasdYield") == 0) {
+            XDodge::SetWasdYieldEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeLateralPref") == 0) {
+            XDodge::SetLateralPrefEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeGoalSticky") == 0) {
+            XDodge::SetGoalStickyEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "xdodgeDangerPenalty") == 0) {
+            XDodge::SetDangerWeight(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "xdodgeStayPenalty") == 0 ||
+                   strcmp(keyBuf, "xdodgeFutureSample") == 0 ||
+                   strcmp(keyBuf, "xdodgeFutureHorizon") == 0 ||
+                   strcmp(keyBuf, "xdodgeFutureStride") == 0) {
+            /* no-op: not supported by current XDodge */
+        // ── RolloutDodge settings ────────────────────────────────────────────
+        } else if (strcmp(keyBuf, "rolloutHorizonTicks") == 0) {
+            RolloutDodge::SetHorizonTicks(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "rolloutSampleStepMs") == 0) {
+            RolloutDodge::SetSampleStepMs(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "rolloutHeadings") == 0) {
+            RolloutDodge::SetHeadingCount(atoi(valueNorm));
+        } else if (strcmp(keyBuf, "rolloutHitScale") == 0) {
+            RolloutDodge::SetHitScale(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "rolloutIntentWeight") == 0) {
+            RolloutDodge::SetIntentWeight(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "rolloutRebuildN") == 0) {
+            RolloutDodge::SetRebuildN(atoi(valueNorm));
+        } else if (strcmp(keyBuf, "rolloutForceBrute") == 0) {
+            RolloutDodge::SetForceBruteForce(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "rolloutAvoidEnemies") == 0) {
+            RolloutDodge::SetAvoidEnemiesEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "rolloutWasdYield") == 0) {
+            RolloutDodge::SetWasdYieldEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "rolloutCommitDwell") == 0) {
+            RolloutDodge::SetCommitDwellEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "rolloutDrawPath") == 0) {
+            RolloutDodge::SetDrawPathEnabled(atoi(valueNorm) != 0);
+        } else if (strcmp(keyBuf, "gameHitboxMult") == 0) {
+            const float v = static_cast<float>(atof(valueNorm));
+            TestTAB::SetGameHitboxOverride(v < 1.0f - 1e-4f, v);
+        } else if (strcmp(keyBuf, "speedHackMult") == 0) {
+            SpeedHack::SetMultiplier(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "autoAbilityEnabled") == 0) {
+            IpcBridge_SetAutoAbilityEnabled(JsonGetBool(json, "value"));
+        } else if (strcmp(keyBuf, "autoAbilityMpPct") == 0) {
+            IpcBridge_SetAutoAbilityMpPct(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "autoAbilityWizardMode") == 0) {
+            IpcBridge_SetAutoAbilityWizardMode(atoi(valueNorm));
+        } else if (strcmp(keyBuf, "playerNoclipActive") == 0) {
+            s_featPlayerNoclipActive.store(JsonGetBool(json, "value") ? 1 : 0, std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "playerNoclipEnabled") == 0) {
+            s_featPlayerNoclipEnabled.store(JsonGetBool(json, "value") ? 1 : 0, std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "playerNoclipHotkey") == 0) {
+            s_featPlayerNoclipHotkeyVk.store(ResolveHotkeyVk(valueNorm), std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "socketHotkeyActive") == 0) {
+            s_featSocketHotkeyActive.store(JsonGetBool(json, "value") ? 1 : 0, std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "socketHotkey") == 0) {
+            s_featSocketHotkeyVk.store(ResolveHotkeyVk(valueNorm), std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "walkTargetX") == 0) {
+            s_featWalkTargetX.store(static_cast<float>(atof(valueNorm)), std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "walkTargetY") == 0) {
+            s_featWalkTargetY.store(static_cast<float>(atof(valueNorm)), std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "walkTargetActive") == 0) {
+            s_featWalkTargetActive.store(JsonGetBool(json, "value") ? 1 : 0, std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "cameraZoomActive") == 0) {
+            IpcBridge_SetCameraZoom(JsonGetBool(json, "value"), s_featCameraZoomValue.load(std::memory_order_relaxed));
+        } else if (strcmp(keyBuf, "cameraZoomValue") == 0) {
+            IpcBridge_SetCameraZoom(s_featCameraZoomActive.load(std::memory_order_relaxed) != 0, static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "cameraAngleActive") == 0) {
+            IpcBridge_SetCameraAngle(JsonGetBool(json, "value"), s_featCameraAngleValue.load(std::memory_order_relaxed));
+        } else if (strcmp(keyBuf, "cameraAngleValue") == 0) {
+            IpcBridge_SetCameraAngle(s_featCameraAngleActive.load(std::memory_order_relaxed) != 0, atoi(valueNorm));
+        } else if (strcmp(keyBuf, "cameraCenteringActive") == 0) {
+            IpcBridge_SetCameraCentering(JsonGetBool(json, "value"), s_featCameraCentered.load(std::memory_order_relaxed) != 0);
+        } else if (strcmp(keyBuf, "cameraCentered") == 0) {
+            IpcBridge_SetCameraCentering(s_featCameraCenteringActive.load(std::memory_order_relaxed) != 0, JsonGetBool(json, "value"));
+        } else if (strcmp(keyBuf, "o3ShieldActive") == 0) {
+            s_featO3ShieldActive.store(JsonGetBool(json, "value") ? 1 : 0, std::memory_order_relaxed);
+        } else if (strcmp(keyBuf, "skinOverrideEnabled") == 0) {
+            IpcBridge_SetSkinOverride(JsonGetBool(json, "value"), s_featSkinOverrideId.load(std::memory_order_relaxed));
+        } else if (strcmp(keyBuf, "skinOverrideId") == 0) {
+            IpcBridge_SetSkinOverride(s_featSkinOverrideEnabled.load(std::memory_order_relaxed) != 0, atoi(valueNorm));
+        } else if (strcmp(keyBuf, "dodgeWasdLookahead") == 0) {
+            DangerPlanner::SetWasdLookahead(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "dodgeTightLeash") == 0) {
+            DangerPlanner::SetTightLeashRadius(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "dodgeIdleMinGain") == 0) {
+            DangerPlanner::SetIdleMinGain(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "dodgeStickiness") == 0) {
+            DangerPlanner::SetStickiness(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "dodgeReplanOnSpawn") == 0) {
+            DangerPlanner::SetReplanOnHazardSpawn(JsonGetBool(json, "value"));
+        } else if (strcmp(keyBuf, "dodgeStrategicBias") == 0) {
+            DangerPlanner::SetStrategicBiasEnabled(JsonGetBool(json, "value"));
+        } else if (strcmp(keyBuf, "dodgeStrategicNearWaypoint") == 0) {
+            DangerPlanner::SetStrategicUseNearWaypoint(JsonGetBool(json, "value"));
+        } else if (strcmp(keyBuf, "dodgeHitAversion") == 0) {
+            DangerPlanner::SetHitAversion(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "dodgeHitScale") == 0) {
+            DangerPlanner::SetDodgeHitScale(static_cast<float>(atof(valueNorm)));
+        } else if (strcmp(keyBuf, "targetFrameRate") == 0) {
+            FpsSetter::SetTargetFps(atoi(valueNorm));
+        }
+    }
+}
+
+// ── Bridge thread ─────────────────────────────────────────────────────────────
+
+DWORD WINAPI IpcBridgeThread(LPVOID)
+{
+    DBG_FILE_LOG("[IpcBridgeThread] Entered (DLL-as-client mode).");
+    DbgLog("Thread started.");
+
+    DBG_FILE_LOG("[IpcBridgeThread] Connecting to pipe: " << PipeName());
+
+    while (!s_shutdown)
+    {
+        if (!WaitNamedPipeA(PipeName(), 2000))
+        {
+            if (s_shutdown) break;
+            Sleep(500);
+            continue;
+        }
+
+        HANDLE hPipe = CreateFileA(
+            PipeName(),
+            GENERIC_READ | GENERIC_WRITE,
+            0, NULL, OPEN_EXISTING, FILE_ATTRIBUTE_NORMAL, NULL);
+
+        if (hPipe == INVALID_HANDLE_VALUE)
+        {
+            const DWORD err = GetLastError();
+            DBG_FILE_LOG("[IpcBridgeThread] CreateFile failed: " << err);
+            if (err == ERROR_PIPE_BUSY)
+                WaitNamedPipeA(PipeName(), 2000);
+            else
+                Sleep(2000);
+            continue;
+        }
+
+        DWORD pipeMode = PIPE_READMODE_BYTE;
+        SetNamedPipeHandleState(hPipe, &pipeMode, NULL, NULL);
+
+        DBG_FILE_LOG("[IpcBridgeThread] Connected. Sending hello...");
+        DbgLog("Connected. Sending hello...");
+
+        Handshake::ResetAuthState(&s_auth);
+
+        char msgBuf[PIPE_BUFFER_SIZE];
+        char helloChallenge[65] = {};
+        if (!Handshake::GenerateChallenge(helloChallenge))
+        {
+            DbgLog("Failed to generate hello challenge.");
+            CloseHandle(hPipe);
+            Sleep(1000);
+            continue;
+        }
+        strncpy_s(s_auth.pendingChallenge, sizeof(s_auth.pendingChallenge), helloChallenge, _TRUNCATE);
+
+        int len = BuildHelloJson(msgBuf, sizeof(msgBuf), helloChallenge);
+        if (!PipeWriteMessage(hPipe, msgBuf, len))
+        {
+            DbgLog("Failed to send hello.");
+            CloseHandle(hPipe);
+            Sleep(1000);
+            continue;
+        }
+
+        char readBuf[PIPE_BUFFER_SIZE];
+        bool authOk = false;
+        ULONGLONG authDeadline = GetTickCount64() + 5000;
+
+        while (GetTickCount64() < authDeadline && !s_shutdown)
+        {
+            int readLen = PipeReadMessage(hPipe, readBuf, sizeof(readBuf) - 1);
+            if (readLen < 0) break;
+            if (readLen > 0)
+            {
+                readBuf[readLen] = '\0';
+                if (DispatchAuthMessage(readBuf, hPipe, msgBuf, sizeof(msgBuf)))
+                {
+                    authOk = s_auth.authenticated;
+                    break;
+                }
+            }
+            Sleep(25);
+        }
+
+        if (!authOk)
+        {
+            DbgLog("Auth failed or timed out. Retrying.");
+            Handshake::ResetAuthState(&s_auth);
+            CloseHandle(hPipe);
+            Sleep(2000);
+            continue;
+        }
+
+        DbgLog("Authenticated. userId=%s", s_auth.userId);
+
+        ULONGLONG lastPlayerPush = 0;
+        s_auth.lastHeartbeatSent = GetTickCount64();
+        s_auth.lastHeartbeatRecv = GetTickCount64();
+        bool connected = true;
+        bool sentUnresolvedClasses = false;
+
+        while (connected && !s_shutdown)
+        {
+            int readLen = PipeReadMessage(hPipe, readBuf, sizeof(readBuf) - 1);
+            if (readLen < 0)
+            {
+                DbgLog("Server disconnected.");
+                connected = false;
+                break;
+            }
+            else if (readLen > 0)
+            {
+                readBuf[readLen] = '\0';
+                if (!DispatchAuthMessage(readBuf, hPipe, msgBuf, sizeof(msgBuf)))
+                {
+                    if (Handshake::IsHealthy(&s_auth))
+                        DispatchCommand(readBuf);
+                }
+            }
+
+            ULONGLONG now = GetTickCount64();
+
+            if (now - s_auth.lastHeartbeatSent >= Handshake::HEARTBEAT_INTERVAL_MS)
+            {
+                if (s_auth.challengePending)
+                {
+                    s_auth.heartbeatMisses++;
+                    DbgLog("Heartbeat miss #%d.", s_auth.heartbeatMisses);
+                    if (s_auth.heartbeatMisses >= Handshake::HEARTBEAT_MAX_MISSES)
+                    {
+                        DbgLog("Too many misses — disconnecting.");
+                        connected = false;
+                        break;
+                    }
+                }
+
+                char nonce[65] = {};
+                if (Handshake::GenerateChallenge(nonce))
+                {
+                    strncpy_s(s_auth.pendingChallenge, sizeof(s_auth.pendingChallenge), nonce, _TRUNCATE);
+                    s_auth.challengePending  = true;
+                    s_auth.lastHeartbeatSent = now;
+                    const uint64_t outSeq = s_auth.nextServerSeq++;
+                    char outMac[65] = {};
+                    if (!ComputeSessionMacHex(s_auth.sessionKey, outSeq, "heartbeat", nonce, outMac)) {
+                        connected = false;
+                        break;
+                    }
+                    len = BuildHeartbeatJson(msgBuf, sizeof(msgBuf), nonce, outSeq, outMac);
+                    if (!PipeWriteMessage(hPipe, msgBuf, len)) { connected = false; break; }
+                }
+            }
+
+            if (Handshake::IsHealthy(&s_auth) && now - lastPlayerPush >= 200)
+            {
+                lastPlayerPush = now;
+                const uint64_t outSeq = s_auth.nextServerSeq++;
+                char payload[256] = {};
+                BuildPlayerSigPayload(payload, sizeof(payload));
+                char outMac[65] = {};
+                if (!ComputeSessionMacHex(s_auth.sessionKey, outSeq, "player", payload, outMac)) {
+                    connected = false;
+                    break;
+                }
+                len = BuildPlayerJson(msgBuf, sizeof(msgBuf), outSeq, outMac);
+                if (!PipeWriteMessage(hPipe, msgBuf, len)) { connected = false; break; }
+            }
+
+            if (Handshake::IsHealthy(&s_auth))
+            {
+                if (IpcBridge_PollSocketHotkeyEvent())
+                {
+                    if (!WriteSignedHotkeyEvent(hPipe, msgBuf, sizeof(msgBuf), "socket", "toggle", true)) {
+                        connected = false;
+                        break;
+                    }
+                }
+
+                const int noclipEnabled = IpcBridge_DrainPendingNoclipEnabled();
+                if (noclipEnabled >= 0)
+                {
+                    if (!WriteSignedHotkeyEvent(
+                        hPipe, msgBuf, sizeof(msgBuf),
+                        "player-noclip", "noclipEnabled",
+                        noclipEnabled != 0))
+                    {
+                        connected = false;
+                        break;
+                    }
+                }
+
+                std::vector<PendingEvent> drained;
+                {
+                    std::lock_guard<std::mutex> lk(s_pendingEventsMutex);
+                    if (!s_pendingEvents.empty()) drained.swap(s_pendingEvents);
+                }
+                for (const auto& ev : drained)
+                {
+                    if (!WriteSignedHotkeyEvent(
+                        hPipe, msgBuf, sizeof(msgBuf),
+                        ev.pluginId, ev.action, true))
+                    {
+                        connected = false;
+                        break;
+                    }
+                }
+                if (!connected) break;
+            }
+
+            if (!sentUnresolvedClasses && RuntimeOffsets::HasGivenUp()) {
+                sentUnresolvedClasses = true;
+                const char* classes = RuntimeOffsets::GetUnresolvedClassNames();
+                if (classes && classes[0] != '\0') {
+                    const uint64_t outSeq = s_auth.nextServerSeq++;
+                    char outMac[65] = {};
+                    if (ComputeSessionMacHex(s_auth.sessionKey, outSeq, "unresolvedClasses", classes, outMac)) {
+                        len = BuildUnresolvedClassesJson(msgBuf, sizeof(msgBuf), classes, outSeq, outMac);
+                        PipeWriteMessage(hPipe, msgBuf, len);
+                    }
+                }
+            }
+
+            Sleep(25);
+        }
+
+        Handshake::ResetAuthState(&s_auth);
+        CloseHandle(hPipe);
+        DbgLog("Disconnected. Will reconnect.");
+    }
+
+    DbgLog("Thread exiting.");
+    Handshake::ClearSharedKeyCache();
+    return 0;
+}
